Removed client.rs Client is now in lib.rs
