+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod auth;
-pub mod client;
 pub mod comments;
 pub mod error;
 pub mod message;
@@ -8,191 +7,245 @@ pub mod submission;
 pub mod subreddit;
 pub mod user;
 pub mod utils;
+use log::trace;
+use std::fmt;
+use std::fmt::{Display, Formatter};
+use std::str::FromStr;
+use std::sync::Arc;
 
-#[cfg(test)]
-mod tests {
+use reqwest::header::HeaderMap;
+use reqwest::{Body, Client as ReqwestClient, ClientBuilder, Response};
+use serde::de::DeserializeOwned;
+use serde::{de, Deserialize, Deserializer};
 
-    use std::str::FromStr;
+use crate::auth::{Authenticator, PasswordAuthenticator};
+use crate::error::http_error::IntoResult;
+use crate::error::internal_error::InternalError;
+use crate::error::Error;
+use crate::message::Inbox;
+use crate::subreddit::response::Subreddits;
+use crate::subreddit::Subreddit;
+use crate::user::me::Me;
+use crate::user::response::{MeResponse, Users};
+use crate::user::User;
+use crate::utils::options::FeedOption;
+use tokio::sync::{RwLock, RwLockReadGuard};
 
-    use serde_json::Value;
+/// This is who you are. This is your identity and you access point to the Reddit API
+#[derive(Clone)]
+pub struct Client<A: Authenticator> {
+    auth: Arc<RwLock<A>>,
+    client: ReqwestClient,
+    user_agent: String,
+    pub oauth: bool,
+}
 
-    use crate::auth::{AnonymousAuthenticator, PasswordAuthenticator};
-    use crate::client::{Client, FullName};
-    use crate::comments::CommentRetriever;
-    use crate::responses::RedditType;
-    use crate::responses::RedditType::{Comment, Link};
-    use crate::submission::{SubmissionRetriever, SubmissionType};
-    use crate::utils::options::FriendType;
-
-    async fn create_logged_in_client() -> Client<PasswordAuthenticator> {
-        dotenv::dotenv().ok();
-        let arc = PasswordAuthenticator::new(
-            std::env::var("CLIENT_KEY").unwrap().as_str(),
-            std::env::var("CLIENT_SECRET").unwrap().as_str(),
-            std::env::var("REDDIT_USER").unwrap().as_str(),
-            std::env::var("PASSWORD").unwrap().as_str(),
-        );
-        return Client::login(arc, "async_rawr test (by u/KingTuxWH)".to_string())
-            .await
-            .unwrap();
+impl<A: Authenticator> Client<A> {
+    /// Logs into Reddit and Returns a Me
+    pub async fn login(auth: Arc<RwLock<A>>, user_agent: String) -> Result<Client<A>, Error> {
+        let client = ClientBuilder::new()
+            .user_agent(user_agent.clone())
+            .build()?;
+        let arc = auth.clone();
+        let mut guard = arc.write().await;
+        let b = guard.oauth();
+        let _x = guard.login(&client, &user_agent).await;
+        Ok(Client {
+            auth,
+            client,
+            user_agent,
+            oauth: b,
+        })
     }
-
-    async fn create_anon_client() -> Client<AnonymousAuthenticator> {
-        return Client::login(
-            AnonymousAuthenticator::new(),
-            "async_rawr test (by u/KingTuxWH)".to_string(),
-        )
-        .await
-        .unwrap();
+    /// Gets the authenticator. Internal use
+    pub async fn get_authenticator(&self) -> RwLockReadGuard<'_, A> {
+        self.auth.read().await
     }
-
-    #[tokio::test]
-    async fn anon_subreddit_tests() {
-        let me = create_anon_client().await;
-        let subreddit = me.subreddit("memes".to_string());
-        let x = subreddit.about().await;
-        let subreddit = x.unwrap();
-        println!("{}", subreddit.data.title.unwrap());
-    }
-
-    #[tokio::test]
-    async fn anon_submissions_test() {
-        let me = create_anon_client().await;
-        let subreddit = me.subreddit("rust".to_string());
-        let submissions = subreddit.get_submissions("hot", None).await.unwrap();
-        let listing = submissions.data;
-        let vec = listing.children;
-        let submission = vec.get(0).unwrap().data.to_submission(&me);
-        let comment_response = submission.get_comments(None).await.unwrap();
-        let comments = comment_response.get(1).unwrap();
-        for x in comments.data.children.iter() {
-            println!("{:?}", x.data);
+    /// Creates a subreddit object. However, this will not tell you if the user exists.
+    pub fn subreddit<T: Into<String>>(&self, name: T) -> Subreddit<A> {
+        Subreddit {
+            me: self,
+            name: name.into(),
         }
     }
 
-    #[tokio::test]
-    async fn post_test() {
-        let me = create_anon_client().await;
-        let _subreddit = me.subreddit("rust".to_string());
-    }
-
-    #[ignore]
-    #[tokio::test]
-    async fn test() {
-        let me = create_logged_in_client().await;
-        let response = me.get("/api/v1/me", true).await.unwrap();
-        println!("{}", response.text().await.unwrap())
-    }
-
-    #[ignore]
-    #[tokio::test]
-    async fn test_inbox() {
-        let me = create_logged_in_client().await;
-
-        let inbox = me.inbox();
-        for x in inbox.get_messages(None, None).await.unwrap().data.children {
-            match x.data {
-                Comment(c) => {
-                    println!("Comment {:?}", c.name);
-                }
-                RedditType::Account(_a) => {}
-                Link(_l) => {}
-                RedditType::Message(m) => {
-                    println!("Message {:?}", m.name);
-                }
-                RedditType::Subreddit(_s) => {}
-                RedditType::Award => {}
-                RedditType::Listing(listing) => {
-                    println!("Listing {:?}", listing.children.len());
-                }
-            }
+    /// Creates a user object. However, this will not tell you if the user exists.
+    pub fn user<T: Into<String>>(&self, name: T) -> User<A> {
+        User {
+            me: self,
+            name: name.into(),
         }
     }
-
-    #[ignore]
-    #[tokio::test]
-    async fn hidden_sub() {
-        let me = create_logged_in_client().await;
-
-        let _response = me
-            .subreddit("RedditNobility".to_string())
-            .about()
-            .await
-            .unwrap();
-    }
-
-    #[ignore]
-    #[tokio::test]
-    async fn friend() {
-        let me = create_logged_in_client().await;
-
-        let _response = me
-            .subreddit("RedditNobility".to_string())
-            .add_friend("LordPenguin42".to_string(), FriendType::Contributor)
-            .await
-            .unwrap();
-    }
-
-    #[ignore]
-    #[tokio::test]
-    async fn test_send() {
-        let me = create_logged_in_client().await;
-
-        let inbox = me.inbox();
-        let result = inbox.compose("LordPenguin42".to_string(),
-                                   "Test from Async Rawr".to_string(),
-                                   "I don’t want to talk to you no more, you empty-headed animal-food-trough wiper. I fart in your general direction. Your mother was a hamster, and your father smelt of elderberries.".to_string(),
-                                   Some("new_rawr".to_string())).await;
-        my_loop(result.unwrap().as_object().unwrap());
-    }
-
-    #[ignore]
-    #[tokio::test]
-    async fn test_block() {
-        let me = create_logged_in_client().await;
-
-        let inbox = me.inbox();
-        inbox
-            .block_author(FullName::from_str("t2_a3bjd54v").unwrap())
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn anon_user_tests() {
-        let me = create_anon_client().await;
-
-        let user = me.user("HoodwinkingGnome".to_string());
-        let result = user.about().await;
-        if let Err(error) = result {
-            println!("{}", error);
-            return;
+    /// Makes a get request with Reqwest response
+    pub async fn get(&self, url: &str, oauth: bool) -> Result<Response, Error> {
+        let mut guard = self.get_authenticator().await;
+        if guard.needs_token_refresh() {
+            trace!("Token Expired. Refreshing");
+            drop(guard);
+            self.re_login();
+            guard = self.get_authenticator().await;
         }
-        let response = result.unwrap();
-        let _submissions = user.submissions(None).await.unwrap();
-        let _comments = user.comments(None).await.unwrap();
-        println!("{}", response.data.name);
+        let string = self.build_url(url, oauth, guard.oauth());
+        let mut headers = HeaderMap::new();
+        guard.headers(&mut headers);
+        drop(guard);
+        self.client
+            .get(string)
+            .headers(headers)
+            .send()
+            .await
+            .map_err(|error| Error::InternalError(InternalError::ReqwestError(error)))
+    }
+    /// Makes a post request with Reqwest response
+    pub async fn post(&self, url: &str, oauth: bool, body: Body) -> Result<Response, Error> {
+        let mut guard = self.get_authenticator().await;
+        if guard.needs_token_refresh() {
+            trace!("Token Expired. Refreshing");
+            drop(guard);
+            self.re_login();
+            guard = self.get_authenticator().await;
+        }
+        let string = self.build_url(url, oauth, guard.oauth());
+        let mut headers = HeaderMap::new();
+        guard.headers(&mut headers);
+        drop(guard);
+        self.client
+            .post(string)
+            .body(body)
+            .headers(headers)
+            .send()
+            .await
+            .map_err(Error::from)
+    }
+    /// Makes a get request with JSON response
+    pub async fn get_json<T: DeserializeOwned>(
+        &self,
+        url: &str,
+        oauth: bool,
+    ) -> crate::error::Result<T> {
+        let response = self.get(url, oauth).await?;
+        response.status().into_result()?;
+        let value = response.text().await?;
+        trace!("{}", &value);
+        let x: T = serde_json::from_str(value.as_str())?;
+        Ok(x)
+    }
+    /// Makes a post request with JSON response
+    pub async fn post_json<T: DeserializeOwned>(
+        &self,
+        url: &str,
+        oauth: bool,
+        body: Body,
+    ) -> crate::error::Result<T> {
+        let response = self.post(url, oauth, body).await?;
+        response.status().into_result()?;
+
+        let value = response.text().await?;
+        trace!("{}", &value);
+        let x: T = serde_json::from_str(value.as_str())?;
+        Ok(x)
+    }
+    /// Builds a URL
+    pub fn build_url(&self, dest: &str, oauth_required: bool, oauth_supported: bool) -> String {
+        let stem = if oauth_required || oauth_supported {
+            // All endpoints support OAuth, but some do not support the regular endpoint. If we are
+            // required to use it or support it, we will use it.
+            assert!(
+                oauth_supported,
+                "OAuth is required to use this endpoint, but your authenticator does not \
+                     support it."
+            );
+            "https://oauth.reddit.com"
+        } else {
+            "https://api.reddit.com"
+        };
+        format!("{stem}{dest}")
     }
 
-    fn my_loop(map: &serde_json::Map<String, Value>) {
-        for x in map {
-            if x.1.is_object() {
-                my_loop(x.1.as_object().unwrap());
-            } else if x.1.is_boolean() {
-                println!("pub {}: Option<bool>,", x.0)
-            } else if x.1.is_null() {
-                println!("pub {}: Option<Value>,", x.0)
-            } else if x.1.is_f64() {
-                println!("pub {}: Option<f64>,", x.0)
-            } else if x.1.is_i64() {
-                println!("pub {}: Option<i64>,", x.0)
-            } else if x.1.is_u64() {
-                println!("pub {}: Option<u64>,", x.0)
-            } else if x.1.is_string() {
-                println!("pub {}: Option<String>,", x.0)
-            } else if x.1.is_number() {
-                println!("pub {}: Option<i64>,", x.0)
-            }
+    /// Searches Reddit for subreddits
+    pub async fn search_subreddits(
+        &self,
+        name: String,
+        limit: Option<u64>,
+        feed: Option<FeedOption>,
+    ) -> crate::error::Result<Subreddits> {
+        let mut url = format!("/subreddits/search?q={name}");
+        if let Some(options) = feed {
+            url.push_str(options.url().as_str());
         }
+        if let Some(limit) = limit {
+            url.push_str(&format!("&limit={limit}"));
+        }
+        self.get_json::<Subreddits>(&url, false).await
+    }
+
+    /// Searches Reddit for Users
+    pub async fn search_users(
+        &self,
+        name: String,
+        limit: Option<u64>,
+        feed: Option<FeedOption>,
+    ) -> crate::error::Result<Users> {
+        let mut url = format!("/users/search?q={name}");
+        if let Some(options) = feed {
+            url.push_str(options.url().as_str());
+        }
+        if let Some(limit) = limit {
+            url.push_str(&format!("&limit={limit}"));
+        }
+        self.get_json::<Users>(&url, false).await
+    }
+    async fn re_login(&self) {
+        let mut guard = self.auth.write().await;
+        guard.login(&self.client, &self.user_agent).await;
+    }
+}
+impl Client<PasswordAuthenticator> {
+    /// Inbox
+    pub fn inbox(&self) -> Inbox {
+        Inbox { me: self }
+    }
+
+    /// Get Me
+    #[allow(clippy::needless_lifetimes)]
+    pub async fn me<'a>(&'a self) -> Result<Me<'a>, Error> {
+        let me: MeResponse = self.get_json("/api/v1/me", true).await?;
+        Ok(Me { client: self, me })
+    }
+}
+pub struct FullName {
+    pub reddit_type: String,
+    pub id: String,
+}
+
+impl<'de> Deserialize<'de> for FullName {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s = String::deserialize(deserializer)?;
+        FullName::from_str(s.as_str()).map_err(de::Error::custom)
+    }
+}
+
+impl FromStr for FullName {
+    type Err = Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let split = s.split('_').collect::<Vec<&str>>();
+        if split.len() == 1 {
+            // Yes, it is always a good time to make a monty python joke.
+            return Err(Error::from("Then shalt thou count to two, no more, no less. Two shall be the number thou shalt count, and the number of the counting shall be two."));
+        }
+        return Ok(FullName {
+            reddit_type: split.get(0).unwrap().to_string(),
+            id: split.get(1).unwrap().to_string(),
+        });
+    }
+}
+
+impl Display for FullName {
+    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+        write!(f, "{}_{}", self.reddit_type, self.id)
     }
 }
