+++ src/submission/mod.rs
@@ -1,11 +1,11 @@
 pub mod response;
 
 use crate::auth::Authenticator;
-use crate::client::Client;
 use crate::comments::CommentRetriever;
 use crate::responses::{GenericListing, ListingArray};
 use crate::submission::response::SubmissionsResponse;
 use crate::utils::options::{CommentOption, FeedOption};
+use crate::Client;
 use async_trait::async_trait;
 
 use crate::error::Error;
