+++ src/user/mod.rs
@@ -2,9 +2,9 @@ pub mod me;
 pub mod response;
 
 use crate::auth::Authenticator;
-use crate::client::Client;
 use crate::comments::response::CommentsResponse;
 use crate::error::Error;
+use crate::Client;
 
 use crate::responses::RedditListing;
 use crate::submission::response::SubmissionsResponse;
